Refactor App component to functional, add styles for displays
