+++ src/components/Display.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import '../styles/Display.css';
 
 function Display(props) {
   const { result } = props;
