+++ src/styles/App.css
@@ -1,0 +1,9 @@
+#root {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  align-content: center;
+  justify-content: center;
+  height: 100%;
+  width: 700px;
+}
